requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.4
chardet==4.0.0
colorama>=0.4.3
cryptography==2.9.2
discord.py==1.7.1
docutils==0.15.2
idna==2.5
multidict==5.1.0
rsa==3.4.2
typing-extensions==3.7.4.3
yarl==1.6.3
pymongo
threading_sched
tabulate